# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "API documentation for Laravel application",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/submit-job-application": {
            "post": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Submit job application",
                "operationId": "926528e2f3e62fb88d6ac4c0accb3aed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "jobId": {
                                        "description": "Id of Job",
                                        "type": "integer"
                                    },
                                    "resume": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "coverLetter": {
                                        "description": "Cover letter for job",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Job Application Id",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "user_id": {
                                            "description": "Job Application Applicant Id",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "job_id": {
                                            "description": "Job Id",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "resume": {
                                            "description": "Job Applicant's Resume",
                                            "type": "string",
                                            "example": "laravel Developer"
                                        },
                                        "cover_letter": {
                                            "description": "Job Applicant's Cover Letter",
                                            "type": "string",
                                            "example": "Firefly Tech"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Unsucessful"
                    },
                    "404": {
                        "description": "Job not found"
                    },
                    "400": {
                        "description": "Repeated Action"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/employer/job-applications": {
            "get": {
                "tags": [
                    "Job Applications"
                ],
                "summary": "Employers get all active job submissions for its company",
                "operationId": "a00005cc1315356be237fe0b84a6fa21",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Job Application Id",
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "resume": {
                                            "description": "Job Applicant's Resume",
                                            "type": "file",
                                            "example": "Sample.pdf"
                                        },
                                        "cover_letter": {
                                            "description": "Job Applicant's Coverletter",
                                            "type": "string",
                                            "example": "Develop and maintain software."
                                        },
                                        "user_id": {
                                            "description": "Job Job Applicant's Id",
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "job_id": {
                                            "description": "Job Application's Company Id",
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "is_approved": {
                                            "description": "Job Application's Approval Status",
                                            "type": "boolean",
                                            "example": "0"
                                        },
                                        "is_rejected": {
                                            "description": "Job Application's Rejection Status",
                                            "type": "boolean",
                                            "example": "0"
                                        },
                                        "status": {
                                            "description": "Job Application's Status",
                                            "type": "boolean",
                                            "example": "1"
                                        },
                                        "job_title": {
                                            "description": "Job Title",
                                            "type": "string",
                                            "example": "Laravel Developer"
                                        },
                                        "company": {
                                            "description": "Job Application Company Name",
                                            "type": "string",
                                            "example": "Firefly Tech"
                                        },
                                        "company_email": {
                                            "description": "Job Application Company Email",
                                            "type": "string",
                                            "example": "company@gmail.com"
                                        },
                                        "applicant_name": {
                                            "description": "Job Application Applicant Name",
                                            "type": "string",
                                            "example": "seeker"
                                        },
                                        "applicant_email": {
                                            "description": "Job Application Applicant Name",
                                            "type": "string",
                                            "example": "applicant@gmail.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}